pyanalyzer/requirements.txt
appdirs==1.4.4
flake8==5.0.4
kafka-python==2.0.2
lxml==4.9.1
mccabe==0.7.0
packaging==21.3
py==1.11.0
py-common-network-task==1.3.4
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pyshark==0.5.3
python-dotenv==0.21.0
